append example
